requirements.txt
tensorflow==2.1.0
Pillow==7.0.0
numpy==1.18.1
scipy==1.4.1